Bluetooth: Audio: Move capability structs to capabilities.h
Remvoe the capability related struct to the proper header file.
